添加配置文件,增加异常捕获
