+++ dir_.bat
@@ -1,0 +1,3 @@
+call activate myth
+pyinstaller -D run_steel_detect.py
+pause
