+++ exe.bat
@@ -1,0 +1,3 @@
+call activate myth
+pyinstaller -F run_steel_detect.py
+pause
